requirements.txt
streamlit
pandas
numpy
plotly
wordcloud
matplotlib
fpdf2
kaleido
